Conflicts:
	.gitignore
	build/web/WEB-INF/classes/Hwp07/Credentials.sql
	build/web/WEB-INF/classes/Hwp07/Customer.class
	build/web/WEB-INF/classes/Hwp07/RegisteredBean.class
	build/web/customerFolder/profile.xhtml
	build/web/index.xhtml
	build/web/regForm.xhtml
	nbproject/private/private.properties
	nbproject/private/private.xml
	src/java/Hwp07/Credentials.sql
	src/java/Hwp07/Customer.java
	src/java/Hwp07/RegisteredBean.java
	web/index.xhtml

+++ src/java/JavaObjects/AddAlbumBean.java
@@ -80,7 +80,7 @@ public class AddAlbumBean implements Serializable {
 
             while (result.next()) {
                 Album a = new Album();
-                a.setAlbumID(result.getLong("ALBUM_ID"));
+                a.setAlbumID(result.getInt("ALBUM_ID"));
                 a.setTitle(result.getString("TITLE"));
                 a.setArtist(result.getString("ARTIST"));
                 a.setReleaseYear(result.getInt("YEAR"));
